refactor: move handlers out from `server` to `src` to keep their code visibly separated
