+++ src/server/extensions.ts
@@ -25,7 +25,7 @@
  ******/
 
 import { Server } from '@hapi/hapi'
-import onPreHandler from './handlers/shared/onPreHandler'
+import onPreHandler from '~/handlers/shared/onPreHandler'
 
 async function register (server: Server): Promise<Server> {
   await server.ext([
